+++ src/AuditableDomainEntity/Interfaces/IDomainEvent.cs
@@ -4,6 +4,6 @@ public interface IDomainEvent
 {
     public Ulid EventId { get; init; }
     public Ulid EntityId { get; init; }
-    public int EventVersion { get; init; }
+    public float EventVersion { get; init; }
     public DateTimeOffset CreatedAtUtc { get; init; }
 }
